refactor: make game settings more extensible
